feat(arr): split code and check for tags
